+++ core/arch/arm/plat-imx/imx-common.c
@@ -86,22 +86,3 @@ bool soc_is_imx7ds(void)
 	return imx_soc_type() == SOC_MX7D;
 }
 
-uint32_t imx_get_src_gpr(int cpu)
-{
-	vaddr_t va = core_mmu_get_va(SRC_BASE, MEM_AREA_IO_SEC);
-
-	if (soc_is_imx7ds())
-		return io_read32(va + SRC_GPR1_MX7 + cpu * 8 + 4);
-	else
-		return io_read32(va + SRC_GPR1 + cpu * 8 + 4);
-}
-
-void imx_set_src_gpr(int cpu, uint32_t val)
-{
-	vaddr_t va = core_mmu_get_va(SRC_BASE, MEM_AREA_IO_SEC);
-
-	if (soc_is_imx7ds())
-		io_write32(va + SRC_GPR1_MX7 + cpu * 8 + 4, val);
-	else
-		io_write32(va + SRC_GPR1 + cpu * 8 + 4, val);
-}
